Remove another un-locked access to activity func map
This is the only access outside internal_worker.go, so hopefully it's the last.
Regardless, this may be a fairly good argument for breaking up this single giant package,
if only to protect internal fields from un-guarded access.  Or for switching to something
that mandates using a lock to access the data (e.g. similar to uber/atomic, but for locks).